# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Abit service.",
        "title": "Abit IS",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/interview": {
            "get": {
                "description": "get a list of all interviews",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interview"
                ],
                "summary": "List all interviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Interview"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new interview to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interview"
                ],
                "summary": "Create a new interview",
                "parameters": [
                    {
                        "description": "Add interview",
                        "name": "interview",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InterviewData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/interview/{id}": {
            "get": {
                "description": "Get details of an interview by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interview"
                ],
                "summary": "Get details of an interview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interview ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Interview"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing interview by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interview"
                ],
                "summary": "Update an existing interview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interview ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Interview data",
                        "name": "semester",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InterviewData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Interview"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a interview by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interview"
                ],
                "summary": "Delete a interview",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interview ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Interview": {
            "type": "object",
            "properties": {
                "interview_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.InterviewData": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}